pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-prompt-evaluation"
version = "0.1.0"
description = "Feature-based prompt evaluation pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "anthropic==0.83.0",
  "python-dotenv==1.2.1",
]

[project.scripts]
evaluate = "main:main"
new = "bootstrap_feature:main"

[tool.setuptools]
py-modules = ["main", "bootstrap_feature"]
